Devise authentication with JWT
It initially (though it's hardly been tested) appears as though authentication with devise and JWT is working.
A lot more testing needs to be done, and I still have a few uncertain points about the setup. Summary:
- Created #2 to remind myself to take a look in test.rb at the mailer config. Devise was complaining about it, so
I added the config for localhost, but I want to make sure that this doesn't cause test problems.