Config rosters_controller and routes for proper auth
The rosters_controller now expects some sort of authentication, as it uses the current_user
Devise variable in order to determine which rosters to load on the index action, and to determine
if the logged in user is permitted to view certain rosters.

Routes were updated so that rosters are no longer after users, since we're sort of treating
them as a single resource now. It is implied that if you visit `api/v1/rosters` that the user will be the logged in user.