Fix rosters controller, add tests
The rosters controller actually had glaring issues in it that were caught
while trying to setup tests for it.
These issues have been resolved I believe, and the tests that exposed them have been created.

The tests are as of yet incomplete, since they lack a bit of detailed checks.