Add launch configurations and rspec binary
Added the correct launch configurations to run the rspec tests as well as the rails server.
Tested both to be functional.
As part of getting the RSpec one working, I ran `bundle binstubs rspec-core` to get the binstub for it.

Also updated the README with the part about installing the ruby-debug-ide extension.