Improve current_roster method
The current_roster method now renders if the roster hasn't been found, and is
intended to be called in a before_action in controllers.
This allows the controller to ensure that a hunt belonging to the user has been provided,
and if not the action is interrupted by the head render.