Improve participants and rosters controllers
With the changes to the current_roster method, and my new way of doing things
that came with it, I've updated the participants and rosters controllers.
Both are now more reliant on before_action hooks, which means I don't
have to do as much error checking in the actions themselves.
The before_actions act as prerequisited to being able to perform
the actions.