DRY up rendering resources
Added some methods in the application controller that can be reused by the other controllers
when rendering resources, since most of the code was the exact same, minus the resource
name.
While doing this, I also decided to switch the response code for bad requests from 422 to 400, based on
the discussion on this article: https://stackoverflow.com/questions/16133923/400-vs-422-response-to-post-of-data.