Fix yet more license, match, participat and round models
I realized that the way that I was handling associations that I intended to be immutable
wasn't enough, since associations can be made on either end; a match can
be created with two licenses, or a license can be created with a match.
Thus, I added `before_add` callbacks on both sides of associations
that I don't want to change under certain conditions.
Tests are also updated, coming later.