Small tweaks to matchmake, more useful hunt and round
Added more useful functionality to the hunt and round models, such as a method in the hunt
to get the current round and round number, as well as fixing the logic
for auto assign number accordingly in the round.

Also made a job for actually performing the matchmake process, and fixed some glaring
bugs in the constructor for the matchmake object.