Fix stupidity in matches controller, matchmaking
I had some pretty dumb mistakes present in the MatchesController, as
well as a learning experience in that active job has trouble passing
keyword arguments to the job. Thus, I changed to keyword parameters to
simply be a hash and got things working.

Also made the job itself more efficient by eagerly loading participant
data for the licenses, as well as by having a default value of an empty
array for the within and between properties instead of nil.
This was changed after I found a bug where if you had an empty array of
within or between properties (as there would possibly be passed in from
the parameters), then all of the licenses would become nil because of a
bad return value.