Prototype for frontend-provided confirmation url
This isn't a complete prototype, but it demonstrates a method by which I
could pass a URL for confirmations from the frontend to the backend
during the API request, and then have that URL passed from the creation
or update of the model all the way to the mailer method that gets the
url.

I do not think I will be implementing this however, as it poses a
possible security flaw, and may be unnecessarily dynamic for a backend
application that is planned to only ever have one client, and doesn't
currently accept requests from other hosts.