Add first draft of roster permissions
A roster permission system where the relationship between users and
rosters is many-to-many has been created.

There is a major pain in this current system however, since each roster
needs at least one associated owner permission, and all permissions need
an associated roster, so there is a bit of a deadlock.

Currently this still works as long as the roster and the owner
permission for it are created as part of the same transaction, but it's
a finicky system for testing and whatnot.