Finalize permission structure
The permission database structure should now be final.
I may change the way the validations on the roster and on the
permissions work a little bit later on if I find a way to improve it,
since right now there are only a few ways that you can create a
permission and an associated roster such that they will get saved.