Use validations to prevent match edit
I didn't like the clunkiness of the readonly method of preventing match
update, so I've changed it to a validation instead.

Also added a small helper method to the hunt in order to determine the
next match id, so I wouldn't have to be so manual about it in the match.

Made small fix to the hunt model's call to `take(1)` which was
unnecessary.