Improve current policies
The console policy as it is should be ready for deployment with most of
the console models. Those are: matches, hunts, participants and
licenses.
Rosters have their own policy, which inherits from the console policy
and specified that the permission isn't always required. This is because
the create and index actions don't occur in the context of a single
roster, so there would be no associated permission.