Improve permissions in factory
The roster factory will now automatically detect if an owner
permission is supplied in the :permissions argument.
If not, it will look in the `owner_permission` transient attribute, else
generate one.

This is what is now used in the permission factory to build the
associated roster, since as it was, the associated roster would be
created as if the permission was an owner even if it wasn't.

Changes are backwards compatible.