Add custom rescue for Pundit::NotAuthorizedError
I wanted to keep the functionality of simply returning 404 not found
when the user attempts to access something that they aren't permitted to
access, but I also needed the forbidden code for other actions.

So for create, update and destroy actions to which the user isn't permitted,
forbidden is returned, while for other actions, not found it the code.