Use explicit require for permissions in policies
Removed the require_permission option for the ConsolePolicy. It was
seemingly creating some issues with unhandled Pundit:NotAuthorizedError
exceptions, so I've gone with instead explicitly requiring it for the
actions that should.

This is also more flexible for things like the match's matchmake action,
which will be getting the permission via a different route.