Remove open column from matches
The open column for matches was essentially duplicating the state that
is really stored in licenses.
A match is closed when any of its licenses are eliminated. It is open as
long as none of its licenses are eliminated.
Therefore, the column for open has been removed, and replaced with two
methods `open?` and `closed?` to facilitate determining if a match is
open or closed.