Change name of scope in license model
The scope name may have been a little too confusing, since it really
just finds all licenses with a match in the specified round. Thinking of
removing it eventually, since it shouldn't be used, as licenses aren't
necessarily tied to a round via a match (in the case where a license
doesn't have a match but is not eliminated).