Change participants.extras to jsonb
The "extras" column of the "participants" table is now a jsonb column.
I made this change after reading a little bit on the efficiency impact
taht it could have with reading and querying. It seems like the right
fit, especially if we'll be searching on it eventually (even though
pg_search doesn't support jsonb columns).