Fix license bulk creation a little more
There was a problem when you requested to bulk create licenses for a
hunt without specifying the participant_ids. It would try to create a
license for all participants with no license in the hunt, but not
necessarily in the roster, so it would essentially try to create a
license for most of the participants in the database.

Resolved and added the test cases to check for this.