Fix race condition in license creation
As described in #28
(https://gitlab.com/kyle_anderson/react-rails-ts/-/issues/28#note_377872172),
there was a race condition which would allow multiple licenses for the
same participant in the same hunt to be created.
This would happen if they were all created at the same time. Soltuion
is to lock the participant association so that only one license can be
created for a participant at a time.