Improve policy handling
I've been thinking about how we handle policy authorization, and doing
it with forcing the policy_class variable is not great.

In the pundit documentation, it's stated that if you need to provide
your policy with more context, then you should make a class that would
have all the necessary information and pass that to the policy.

I then realized that such a class exists for me: the model itself. I've
therefore decided to simply create a dummy instance of the model for
authorization whenever it's used.