Switch to 422 for some error codes
I've realized again that unprocessable entity is more descriptive than
bad request, and some of my usages of bad request fit the bill of
unprocessable entity better.

According to
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422,
unprocessable entity means the server understood the request
syntactically but logically it didn't make sense, whereas bad request
means there is something wrong with the syntax.
In most of my error cases, there's nothing wrong with the syntax, but
the request just doesn't make sense or is invalid, thus making 422 a
better code choice I think.
(Not that it really matters a whole lot).

I also fixed up some of the participant upload logic so that it fails if
the headers aren't set to the expected value, which lets me
differentiate that from bad data and return a bad request error code
instead of unprocessble entity when the CSV file is malformed.