Fix DATABASE_URL environment variable name
The DATABASE_URL environment variable may override the database
configurations set in database.yml, which was causing my test
environment to use the same database as development, which was bad.

This variable is defined in my .env file, which doesn't get tracked by
git because of the .gitignore, but the README does tell you to configure
this variable.
Thus, the variable name was changed and the README has been updated.