Use active storage in InstantPrintJob
The instant print job will now be using actual storage locations for the
files it generates and the files that it uses.

This has been tested briefly in the rails console, with the template_pdf
set manually. Going to continue with mechanisms for uploading the
template PDF with a controller and viewing the license printout.