Add attachment_url helper method, include in Hunt
Outside of controllers, there isn't an easy method for getting the URL
to an active storage attachment, so I've created my own which accounts
for nil input and whatnot, and uses the hostname and port configured
with the mailer.

I've added a method in the Hunt model which returns the attachment URLs
for all of its attachments using this method.