Return match local_id instead of ID.
For the frontend, it's more helpful if it receives the match local ID
instead of the match UUID, since everything is done relative to the hunt
at that point. Thus, the match_ids method in license has been changed to
match_numbers, and it returns an array of the license's matches' local
IDs.