Fixup round stuff
The controller was a mess of stuff that doesn't work. To get it working
again with the new ideology of creating new model objects for
authorization, I changed the match policy a little so that it would find
the permission directly instead of relying on the `has_one :through`
association. This was done since the association doesn't work when there
are multiple new objects in the chain, for example a new match whose
round object is not saved to the database.

Anyway, long story short, lots of fixes for code that's not under unit
tests.