Add full specs for instant print controller
Full specs to ensure functionality of the instant print controller have
been added.

A lot of the testing done in these specs is really testing the instant
print job. I'm not sure even if it's worth testing the invalid options
on the instant print job again in its own unit test, so I might skip
that and just test its functionality.

During test creation, I realized that the instant print controller
lacked authorization, so I've added that.

I also realized that there needed to be more validations on the input to
the instant print job, so I've added a ton of those, and made constants
for the error messages to make them accessible outside of the job class.