Improve GitLab-CI with JUnit, update gems
The GitLab CI job should now output the test results in JUnit format,
and then they should be displayed nicely in the GitLab UI.

To do this, I had to add the JUnit formatter for RSpec, as a gem.
While I was going about this, I also ran a `bundle update`.