WIP: Change to API only app
Still working on changing to an API only app. Things work when I test
from the browser, however many of the RSpec tests were failing until I
changed `config.api_only` back to `false` in `application.rb`.

I'm asking on StackOverflow and on the Rails community about some of the
issues that I'm having:
https://discuss.rubyonrails.org/t/rails-api-mode-applicationcontroller-ignores-rails-application-config-action-controller-allow-forgery-protection/76942

https://stackoverflow.com/questions/65688634/rails-api-mode-applicationcontroller-ignores-rails-application-config-action-con

The only good news is that in its current state, the tests do pass.