Fix GitLab CI config
Never tested the config when I changed it, so obviously it was wrong.

The --format option to RSpec has to be just before the --out file to
which it applied, so I have reversed the argument orderings.

Additionally, I had the wrong argument of --output when it should have
been --out.