Remove the asset pipeline
One of the issue I had after converting to an API only app was that I
was still making calls to configure the asset pipeline in initializers
like application.rb.

This manifested itself as a problem with not having required sprockets,
however the better solution is just to remove the asset pipeline
altogether, since we aren't using it anyway.

For this reason, the configurations and files in the asset pipeline have
all been removed.