Remove spring requirement from binstubs
We don't use spring in our production environment, it's just supposed to
speed up development.

We were getting errors launching to Heroku because of this. Spring is
supposed to be loaded conditionally, yet the generated binstubs seemed
to absolutely require it.

I've hopefully fixed this by removing the line, but I'm not sure why the
generated binstubs acted in this way.